Correções e atualizações

+++ teste/src/main/java/com/blog/teste/controller/PostController.java
@@ -10,9 +10,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import javax.servlet.http.HttpServletRequest;
 
+import org.apache.catalina.connector.Response;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.validation.BindingResult;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,15 +23,16 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.blog.teste.model.Comentario;
 import com.blog.teste.model.Post;
+import com.blog.teste.model.Usuario;
+import com.blog.teste.payload.response.MessageResponse;
 import com.blog.teste.service.ComentarioService;
 import com.blog.teste.service.PostService;
+import com.blog.teste.service.UsuarioService;
 
 @RestController
 @RequestMapping(value="/api")
@@ -39,6 +43,9 @@ public class PostController {
 	
 	@Autowired
 	ComentarioService comentarioService;
+	
+	@Autowired
+	UsuarioService usuarioService;
 	
 	public static String uploadDirectory = System.getProperty("user.dir") + "/src/main/resources/static/imagens";
 	
@@ -52,9 +59,9 @@ public class PostController {
 		return postService.findById(id);
 	}
 		
-	
+	@CrossOrigin(origins = "http://localhost:8080")
 	@PostMapping(value = "/newpost")
-	public Post savePost(@RequestBody Post post, MultipartFile imagem) throws IOException {
+	public Post savePost(@RequestBody Post post, MultipartFile imagem, HttpServletRequest request) throws IOException {
 			
 		if(imagem != null) {
 		
@@ -72,30 +79,27 @@ public class PostController {
 			post.setImagemNome(imagemNome);
 			post.setImagemTipo(imagemTipo);
 			post.setImagemTamanho(imagemTamanho);
+			post.setAutor(request.getRemoteUser());
 			
 		}
-			
+		
+		post.setAutor(request.getRemoteUser());	
 		post.setData(LocalDate.now());
 		return postService.save(post);
 	}
 	
 	@PutMapping(value="/post/{id}")
-	public Post inserirComentarios(@PathVariable("id") String idPost, @RequestBody Comentario comentario) {
+	public Post inserirComentarios(@PathVariable("id") String idPost, @RequestBody Comentario comentario, HttpServletRequest request) {
 		
 			Optional<Post> optionalPost = postService.findById(idPost);
-			Post post = optionalPost.get();	
-			
-			//comentario.setPost(id);
-				
-			/*
-			 * comentario.stream() .forEach(coment -> coment.setPost(post.getId()));
-			 */		
+			Post post = optionalPost.get();		
 			
 				List<Comentario> coment = new ArrayList<>();
 				
 				coment.add(comentario);
 				
 				post.setComentarios(coment);
+				comentario.setAutor(request.getRemoteUser());
 				comentario.setPost(idPost);
 				
 				comentarioService.save(comentario);
@@ -104,27 +108,47 @@ public class PostController {
 			
 	}
 	
-	
-	 @DeleteMapping("/post") public void deletePost(@RequestBody Post post) { 
+	 @DeleteMapping("/post") 
+	 public ResponseEntity<?> deletePost(@RequestBody Post post, HttpServletRequest request) { 
 		 
 	 List<Comentario> comentarios = new ArrayList<>();
 	 
-	 Optional<Post> optionalPost = postService.findById(post.getId());
-		Post postGet = optionalPost.get();	
+	 String remoteUser = request.getRemoteUser();
 	 
-	 comentarios.addAll(postGet.getComentarios());
+	 Optional<Post> pId = postService.findById(post.getId());
+	 Post p = pId.get();
 	 
-	 comentarioService.deleteAll(comentarios);
-	 
-	 postService.delete(postGet, comentarios, uploadDirectory); 
-	 
+	 if(!remoteUser.equals(p.getAutor())) {
+		 return ResponseEntity.badRequest()
+					.body(new MessageResponse("Somente o autor tem autorização para exclusão"));
+	 }else {
+			 Optional<Post> optionalPost = postService.findById(post.getId());
+			 Post postGet = optionalPost.get();	
+			 
+			 comentarios.addAll(postGet.getComentarios());
+			 
+			 comentarioService.deleteAll(comentarios);
+			 
+			 postService.delete(postGet, comentarios, uploadDirectory);
+			
+			 return ResponseEntity.ok(new MessageResponse("Deletado com sucesso")); 
+	 	}
 	 }
 	 
 	
-	@DeleteMapping("/deletarComentario/")
-	public void deleteComentario(@RequestBody Comentario comentario) {
-		comentario = comentarioService.findById(comentario.getId());
+	@DeleteMapping("/comentario")
+	public ResponseEntity<?> deleteComentario(@RequestBody Comentario comentario, HttpServletRequest request) {
+		
+		 comentario = comentarioService.findById(comentario.getId());
+		 String remoteUser = request.getRemoteUser();
+		 
+		 if(!remoteUser.equals(comentario.getAutor())) {
+			 return ResponseEntity.badRequest()
+						.body(new MessageResponse("Somente o autor tem autorização para exclusão"));
+		 }else {	
 			comentarioService.delete(comentario);
+			return ResponseEntity.ok(new MessageResponse("Deletado com sucesso"));
+		 }
 	}
